pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "p2p"
version = "0.1.0"
description = "Convert PDF slides to editable PPTX using multimodal LLM"
requires-python = ">=3.11"
dependencies = [
    "python-pptx>=0.6.22",
    "PyMuPDF>=1.24.0",
    "Pillow>=10.0.0",
    "openai>=1.30.0",
    "lxml>=5.0.0",
    "tiktoken>=0.7.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff",
    "pylint",
    "mypy",
    "lxml-stubs",
    "pytest",
]

[project.scripts]
p2p = "src.main:main"

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.ruff.format]
line-ending = "lf"

[tool.pylint.format]
max-line-length = 120

[tool.pylint."messages control"]
disable = [
    "missing-module-docstring",
    "missing-class-docstring",
    "missing-function-docstring",
    "too-few-public-methods",
    "import-error",
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
